+++ tests/intergration/test_account_api.py
@@ -5,7 +5,7 @@
 
 from convex_api import (
     Account,
-    ConvexAPI,
+    API,
     KeyPair,
 )
 
@@ -13,15 +13,15 @@ TEST_ACCOUNT_NAME = 'test.convex-api'
 
 def test_account_api_create_account(convex_url):
 
-    convex = ConvexAPI(convex_url)
-    key_pair = KeyPair.create()
+    convex = API(convex_url)
+    key_pair = KeyPair()
     result = convex.create_account(key_pair)
     assert(result)
 
 
 def test_account_api_multi_create_account(convex_url):
-    convex = ConvexAPI(convex_url)
-    key_pair = KeyPair.create()
+    convex = API(convex_url)
+    key_pair = KeyPair()
     account_1 = convex.create_account(key_pair)
     assert(account_1)
     account_2 = convex.create_account(key_pair)
@@ -34,7 +34,7 @@ def test_account_api_multi_create_account(convex_url):
 
 
 def test_account_name(convex_url, test_key_pair_info):
-    convex = ConvexAPI(convex_url)
+    convex = API(convex_url)
     import_key_pair = KeyPair.import_from_bytes(test_key_pair_info['private_bytes'])
     if convex.resolve_account_name(TEST_ACCOUNT_NAME):
         account = convex.load_account(TEST_ACCOUNT_NAME, import_key_pair)
@@ -49,7 +49,7 @@ def test_account_name(convex_url, test_key_pair_info):
 
 
 def test_account_setup_account(convex_url, test_key_pair_info):
-    convex = ConvexAPI(convex_url)
+    convex = API(convex_url)
     import_key_pair = KeyPair.import_from_bytes(test_key_pair_info['private_bytes'])
     account = convex.setup_account(TEST_ACCOUNT_NAME, import_key_pair)
     assert(account.address)