+++ tests/unit/test_key_pair.py
@@ -19,7 +19,7 @@ from convex_api.utils import (
 SIGN_HASH_TEXT = '5bb1ce718241bfec110552b86bb7cccf0d95b8a5f462fbf6dff7c48543622ba5'
 SIGN_TEXT = '0x7eceffab47295be3891ea745838a99102bfaf525ec43632366c7ec3f54db4822b5d581573aecde94c420554f963baebbf412e4304ad8636886ddfa7b1049f70e'
 def test_key_pair_create_new():
-    key_pair = KeyPair.create()
+    key_pair = KeyPair()
     assert(key_pair)
     assert(key_pair.public_key)
 
@@ -81,7 +81,7 @@ def test_key_pair_export_to_mnemonic(test_key_pair):
     assert(test_key_pair.export_to_mnemonic == new_key_pair.export_to_mnemonic)
 
 def test_key_pair_is_equal(test_key_pair):
-    key_pair = KeyPair.create()
+    key_pair = KeyPair()
     assert(test_key_pair.is_equal(test_key_pair))
     assert(test_key_pair.is_equal(test_key_pair.public_key_api))
     assert(test_key_pair.is_equal(test_key_pair.public_key))